Interval addition, win after 10 rounds, solutions generated up front etc.
basically an overhaul, restructured on the interval model where solutions are generated up front, computerPlay continually runs until the number of flashes equals the current level, and then the buttons become clickable by the user. if the click is incorrect, the game is reset. if not, and if the click is the last in the previous sequence, then computerplay continues to run, this time lighting up one new button at the end